rawkv: move rawkv client to rawkv (#227)
